+++ test/runtime/test_module_registry.cpp
@@ -20,7 +20,7 @@ struct TestServiceScope
 
   TestServiceScope()
   {
-    alloc.Init();
+    (void)SetAllocator(&alloc);
     jobs.Init();
     profiler.Init();
     logger.Init();
@@ -28,7 +28,6 @@ struct TestServiceScope
     (void)modules.Init();
 
     Services svc {
-        .Allocator = &alloc,
         .JobSystem = &jobs,
         .Profiler = &profiler,
         .Logger = &logger,
@@ -41,6 +40,7 @@ struct TestServiceScope
   ~TestServiceScope()
   {
     UninstallServices();
+    ResetAllocator();
   }
 };